Externalizing Context Entry props
